Fixed bug at Time Attack Screen
In the case when all levels are unlocked the "next stage to unlock code" errored because there is no stage after 10.

It would only happen the second time getting to the time attack stage.

Now it only runs that code if there is a new stage that could be unlocked.