formatting change and fixed length bug.
line is getting a bit long and fits with the dom element formating.

reordered the .setOrigin and .setDepth to be consistent with the rest of the document.

deleting doesn't delete the array allocation and so pushing will always add to the length of the array and add a new unallocated spot.