Added snake body to show when portalling.
Adds a snake body segment that checks if the second to last piece connects with a portal.

Adds the snake piece when the head passes over the portal
Removes it when the second to last part passes over the portal.

// History
Added extra snake part.

Unfortunately the never touches the portal so the toggle doesn't work.

Stil doesn't wprk