Fixed No Food Spawn Tiles
Most of the code was in commit 2ab34fab0b6bfe7a650a34920cb621476fa4c224 "New level for teaching changing direction while portaling", but I forgot to document the change.

More rebust code for handling the interact layer.

The code tries to run .onOver IFF the attribute exists and is a function.

Also for "No food" spawn tiles to be stored in places for no food to spawn.

No food tiles are stored as a global attribute and checked in the Food Layer. (Note: All of the No Food Tiles should me moved to a Food Layer. This is because sometimes a portal rule tile and a no food spawn tile need to exist in the same spot.)