Fixed Zeds Bonus and made it smaller/
Allowed Zed bonus to support smaller values in the UI.

Make the maximum Zed bonus be %9.9

After level 99 is over leveling. The bonus is a placeholder until we come up with a better way to reward them for getting to 99 but to stop doing it after that level.

Basically a good reason to call it over leveling besides it being a number only soft cap.

Right now it is similar in scale to the other bonuses (No Bonk etc...) even at full level.

Goal is to have it demotivate less high skill playing.