Script for syncing local data to the correct Stage Name
Fixes a bug where if we change the stage name in tiled but keep the UUID it didn't change the local storage entry.

This broke the level unlock chain sometimes until you got a better best of on a stage with an updated name.

Now STAGE_FILE contains all of the active stages in the game and checks on the start screen that all of the local data is updated correctly.