Physics Editor plinko collision
It works most of the time.
I added in phantom balls at the end to help things move through. I think with the pressure of the balls behind it as the zeds mechanism rolls will be sufficient to work smoothly.

I left in the spawner to work every time so we can see how often it works while we test.

commit 26366777472cacc76f9a5260017ac741291fffa8
Author: ltj866 <98800297+ltj866@users.noreply.github.com>
Date:   Tue Nov 19 23:23:15 2024 -0800

    Guiding Happens only on Classic Mode the first time.

commit edb6113c20864bc18d78ea714e352e01fb38f4ba
Author: ltj866 <98800297+ltj866@users.noreply.github.com>
Date:   Tue Nov 19 23:09:47 2024 -0800

    Added collision from the physics editor

    Added 10 invisible balls to go after to keep it cleared.

    Works pretty solidly. I think having a tube above with balls will help the physics.