Added Expert Path Check and some level testing
On expert a black hole will now not spawn unless you have played all the way to the end of a path containing that stage at least once.

It means you can only start a path if you can finish it on expert.

Trying a few things on levels. I think it would help to have an OR portal symbol logic tile.